requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
Pillow
grpcio==1.32.0
numpy
opencv-python
scipy
tensorflow-serving-api